requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2024.2.2=hbcca054_0
cachetools=5.3.1=pypi_0
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.1.0=pypi_0
cryptography=38.0.4=py310h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
debugpy=1.6.7=py310h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
faiss-cpu=1.7.4=py3.10_h8c27c75_0_cpu
falconn=1.3.1=pypi_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
filelock=3.9.0=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.42.1=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.4.0=pypi_0
gast=0.4.0=pypi_0
gdown=4.7.1=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.41=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.57.0=pypi_0
h5py=3.7.0=py310he06866b_0
hdf5=1.10.6=h3ffc7dd_1
idna=3.4=py310h06a4308_0
imageio=2.33.1=pypi_0
imgaug=0.4.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.29.3=pyhd33586a_0
ipython=8.22.2=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.2=py310hff52083_0
kaggle=1.6.0=pypi_0
keras=2.13.1=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libclang=16.0.6=pypi_0
libdeflate=1.17=h5eee18b_1
libfaiss=1.7.4=h2bc3f7f_0_cpu
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.4=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=pypi_0
numpy=1.24.3=py310hd5efca6_0
numpy-base=1.24.3=py310h8e6c178_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=22.0=py310h06a4308_0
pandas=2.0.3=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.0.0=pypi_0
pip=23.2.1=py310h06a4308_0
platformdirs=4.2.0=pyhd8ed1ab_0
pooch=1.4.0=pyhd3eb1b0_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.24.1=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyflann-py3=0.1.0=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py310h06a4308_0
pysparnn=0.4=pypi_0
python=3.10.12=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python-slugify=8.0.1=pypi_0
python_abi=3.10=2_cp310
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py310h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.28.1=py310h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scann=1.2.10=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.0=py310hd5efca6_0
seaborn=0.12.2=pypi_0
sentry-sdk=1.40.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=58.2.0=pypi_0
shapely=2.0.2=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorflow=2.13.0=pypi_0
tensorflow-estimator=2.13.0=pypi_0
tensorflow-io-gcs-filesystem=0.33.0=pypi_0
termcolor=2.3.0=pypi_0
text-unidecode=1.3=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.12.9=pypi_0
tk=8.6.12=h1ccaba5_0
torch=1.12.1+cu102=pypi_0
torchaudio=2.1.2+cu118=pypi_0
torchvision=0.13.1+cu102=pypi_0
tornado=6.1=py310h5764c6d_3
tqdm=4.66.1=pypi_0
traitlets=5.14.2=pyhd8ed1ab_0
triton=2.1.0=pypi_0
typing-extensions=4.5.0=pypi_0
typing_extensions=4.9.0=py310h06a4308_1
tzdata=2023.3=pypi_0
urllib3=1.26.14=py310h06a4308_0
vptree=1.3=pypi_0
wandb=0.16.2=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
werkzeug=2.3.7=pypi_0
wheel=0.38.4=py310h06a4308_0
wrapt=1.15.0=pypi_0
xz=5.4.2=h5eee18b_0
zeromq=4.3.5=h6a678d5_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
